+++ wsi_service/tests/api/test_api_slide_close.py
@@ -4,11 +4,8 @@ import pytest
 import requests_mock
 
 from wsi_service.models.slide import SlideInfo
-from wsi_service.tests.test_api_helpers import (
-    client_changed_timeout,
-    get_image,
-    setup_mock,
-)
+from wsi_service.tests.api.test_api_helpers import client_changed_timeout
+from wsi_service.tests.test_helper import get_image, setup_mock
 
 
 @requests_mock.Mocker(real_http=True, kw="requests_mock")