+++ wsi_service/tests/integration/test_integration_general.py
@@ -1,24 +1,32 @@
-import os
+import requests
 
-from wsi_service.tests.test_api_helpers import (
-    client,
-    client_invalid_data_dir,
-    client_no_data,
+from wsi_service.tests.integration.test_integration_helper import (
+    add_absolute_test_dir_to_env,
+    copy_env,
+    docker_compose_file,
+    wsi_service,
 )
+from wsi_service.tests.test_helper import fetch_test_data
 
 
-def test_get_cases_valid(client):
-    response = client.get("/v1/cases/")
+def test_alive(wsi_service):
+    response = requests.get(f"{wsi_service}/alive")
+    assert response.status_code == 200
+    assert response.json()["status"] == "ok"
+
+
+def test_get_cases_valid(wsi_service):
+    response = requests.get(f"{wsi_service}/v1/cases/")
     assert response.status_code == 200
     cases = response.json()
-    assert len(cases) == 10
+    assert len(cases) == 11
     assert len(cases[0].keys()) == 3
     case = list(filter(lambda case: case["local_case_id"] == "Olympus", cases))[0]
     assert case["case_id"] == "7a32e2c36ca756d9b7df0b627ace4c12"
 
 
-def test_get_available_slides_valid(client):
-    response = client.get("/v1/cases/4593f30c39d75d2385c6c8811c4ae7e0/slides/")
+def test_get_available_slides_valid(wsi_service):
+    response = requests.get(f"{wsi_service}/v1/cases/4593f30c39d75d2385c6c8811c4ae7e0/slides/")
     assert response.status_code == 200
     slides = response.json()
     slide = list(
@@ -36,8 +44,8 @@ def test_get_available_slides_valid(client):
     assert slide["slide_storage"]["storage_addresses"][0]["slide_id"] == "f863c2ef155654b1af0387acc7ebdb60"
 
 
-def test_get_slide_valid(client):
-    response = client.get("/v1/slides/4b0ec5e0ec5e5e05ae9e500857314f20")
+def test_get_slide_valid(wsi_service):
+    response = requests.get(f"{wsi_service}/v1/slides/4b0ec5e0ec5e5e05ae9e500857314f20")
     assert response.status_code == 200
     slide = response.json()
     assert len(slide.keys()) == 3
@@ -51,47 +59,13 @@ def test_get_slide_valid(client):
     assert slide["slide_storage"]["storage_addresses"][0]["slide_id"] == "4b0ec5e0ec5e5e05ae9e500857314f20"
 
 
-def test_get_cases_no_data(client_no_data):
-    response = client_no_data.get("/v1/cases/")
-    assert response.status_code == 200
-    cases = response.json()
-    assert len(cases) == 0
-
-
-def test_get_cases_two_empty_cases(client_no_data):
-    os.mkdir(os.path.join(os.environ["data_dir"], "case0"))
-    os.mkdir(os.path.join(os.environ["data_dir"], "case1"))
-    response = client_no_data.get("/v1/cases/")
-    assert response.status_code == 200
-    cases = response.json()
-    assert len(cases) == 2
-
-
-def test_get_available_slides_empty_case(client_no_data):
-    os.mkdir(os.path.join(os.environ["data_dir"], "case0"))
-    response = client_no_data.get("/v1/cases/")
-    assert response.status_code == 200
-    cases = response.json()
-    case_id = cases[0]["case_id"]
-    response = client_no_data.get(f"/v1/cases/{case_id}/slides/")
-    assert response.status_code == 200
-    slides = response.json()
-    assert len(slides) == 0
-
-
-def test_get_available_slides_invalid_case_id(client):
-    response = client.get("/v1/cases/invalid_id/slides/")
+def test_get_available_slides_invalid_case_id(wsi_service):
+    response = requests.get(f"{wsi_service}/v1/cases/invalid_id/slides/")
     assert response.status_code == 400
     assert response.json()["detail"] == "Case with case_id invalid_id does not exist"
 
 
-def test_get_slide_invalid_slide_id(client):
-    response = client.get("/v1/slides/invalid_id")
+def test_get_slide_invalid_slide_id(wsi_service):
+    response = requests.get(f"{wsi_service}/v1/slides/invalid_id")
     assert response.status_code == 400
     assert response.json()["detail"] == "Slide with slide_id invalid_id does not exist"
-
-
-def test_get_case_invalid_dir(client_invalid_data_dir):
-    response = client_invalid_data_dir.get("/v1/cases/")
-    assert response.status_code == 404
-    assert response.json()["detail"] == "No such directory: /data/non_existing_dir"