+++ wsi_service/tests/integration/test_integration_wsi.py
@@ -1,12 +1,19 @@
 import pytest
+import requests
 import requests_mock
 import tifffile
 from PIL import Image
 
 from wsi_service.models.slide import SlideInfo
 from wsi_service.settings import Settings
-from wsi_service.tests.test_api_helpers import (
-    client,
+from wsi_service.tests.integration.test_integration_helper import (
+    add_absolute_test_dir_to_env,
+    copy_env,
+    docker_compose_file,
+    wsi_service,
+)
+from wsi_service.tests.test_helper import (
+    fetch_test_data,
     get_image,
     get_tiff_image,
     setup_mock,
@@ -26,13 +33,14 @@ from wsi_service.tests.test_api_helpers import (
         ("56ed11a2a9e95f87a1e466cf720ceffa", 5, 8, 6, 498, (512, 512), 24960, 34560),  # ome-tif 5x8bit
         ("cdad4692405c556ca63185bee512e95e", 3, 8, 10, 232, (256, 256), 114943, 76349),  # bif
         ("c4682788c7e85d739ce043b3f6eaff70", 3, 8, 6, 250, (256, 256), 106259, 306939),  # scn
+        ("5c1c0cc5cd3a501480fc6a4cb04ddda8", 3, 8, 9, 250, (128, 128), 192518, 88070),  # isyntax
     ],
 )
 def test_get_slide_info_valid(
-    client, slide_id, channels, channel_depth, num_levels, pixel_size_nm, tile_size, x, y, **kwargs
+    wsi_service, slide_id, channels, channel_depth, num_levels, pixel_size_nm, tile_size, x, y, **kwargs
 ):
     setup_mock(kwargs)
-    response = client.get(f"/v1/slides/{slide_id}/info")
+    response = requests.get(f"{wsi_service}/v1/slides/{slide_id}/info")
     assert response.status_code == 200
     slide_info = SlideInfo.parse_obj(response.json())
     assert slide_info.num_levels == num_levels
@@ -64,10 +72,11 @@ def test_get_slide_info_valid(
         ("56ed11a2a9e95f87a1e466cf720ceffa", 200, (10, 10), (87, 51, 23), (5654, 3341, 1542)),
         ("cdad4692405c556ca63185bee512e95e", 200, (5, 5), (241, 241, 241), (241, 241, 241)),
         ("c4682788c7e85d739ce043b3f6eaff70", 200, (5, 5), (221, 212, 219), (221, 212, 219)),
+        ("5c1c0cc5cd3a501480fc6a4cb04ddda8", 200, (5, 5), (254, 254, 254), (254, 254, 254)),
     ],
 )
 def test_get_slide_thumbnail_valid(
-    client,
+    wsi_service,
     image_format,
     image_quality,
     slide_id,
@@ -80,8 +89,8 @@ def test_get_slide_thumbnail_valid(
     setup_mock(kwargs)
     max_size_x = 21
     max_size_y = 22
-    response = client.get(
-        f"/v1/slides/{slide_id}/thumbnail/max_size/{max_size_x}/{max_size_y}?image_format={image_format}&image_quality={image_quality}",
+    response = requests.get(
+        f"{wsi_service}/v1/slides/{slide_id}/thumbnail/max_size/{max_size_x}/{max_size_y}?image_format={image_format}&image_quality={image_quality}",
         stream=True,
     )
     assert response.status_code == return_value
@@ -120,14 +129,16 @@ def test_get_slide_thumbnail_valid(
         ("56ed11a2a9e95f87a1e466cf720ceffa", False, (), ()),
         ("cdad4692405c556ca63185bee512e95e", False, (), ()),
         ("c4682788c7e85d739ce043b3f6eaff70", False, (), ()),
+        ("5c1c0cc5cd3a501480fc6a4cb04ddda8", True, (200, 200), (255, 255, 255)),
     ],
 )
 def test_get_slide_label_valid(
-    client, image_format, image_quality, slide_id, has_label, pixel_location, testpixel, **kwargs
+    wsi_service, image_format, image_quality, slide_id, has_label, pixel_location, testpixel, **kwargs
 ):
     setup_mock(kwargs)
-    response = client.get(
-        f"/v1/slides/{slide_id}/label?image_format={image_format}&image_quality={image_quality}", stream=True
+    response = requests.get(
+        f"{wsi_service}/v1/slides/{slide_id}/label?image_format={image_format}&image_quality={image_quality}",
+        stream=True,
     )
     if has_label:
         assert response.status_code == 200
@@ -169,14 +180,16 @@ def test_get_slide_label_valid(
         ("56ed11a2a9e95f87a1e466cf720ceffa", 404, (), ()),
         ("cdad4692405c556ca63185bee512e95e", 200, (0, 0), (60, 51, 36)),
         ("c4682788c7e85d739ce043b3f6eaff70", 200, (0, 0), (3, 3, 3)),
+        ("5c1c0cc5cd3a501480fc6a4cb04ddda8", 200, (0, 0), (189, 210, 203)),
     ],
 )
 def test_get_slide_macro_valid(
-    client, image_format, image_quality, slide_id, return_value, pixel_location, testpixel, **kwargs
+    wsi_service, image_format, image_quality, slide_id, return_value, pixel_location, testpixel, **kwargs
 ):
     setup_mock(kwargs)
-    response = client.get(
-        f"/v1/slides/{slide_id}/macro?image_format={image_format}&image_quality={image_quality}", stream=True
+    response = requests.get(
+        f"{wsi_service}/v1/slides/{slide_id}/macro?image_format={image_format}&image_quality={image_quality}",
+        stream=True,
     )
     assert response.status_code == return_value
     if return_value == 200:
@@ -206,17 +219,18 @@ def test_get_slide_macro_valid(
         ("7304006194f8530b9e19df1310a3670f", (0, 0), (231, 182, 212), 50000, 90000, 345),
         ("cdad4692405c556ca63185bee512e95e", (0, 0), (245, 241, 242), 30000, 30000, 345),
         ("c4682788c7e85d739ce043b3f6eaff70", (0, 0), (131, 59, 122), 50000, 55000, 345),
+        ("5c1c0cc5cd3a501480fc6a4cb04ddda8", (0, 0), (234, 234, 247), 140000, 50000, 345),
     ],
 )
 def test_get_slide_region_valid_brightfield(
-    client, image_format, image_quality, slide_id, pixel_location, testpixel, start_x, start_y, size, **kwargs
+    wsi_service, image_format, image_quality, slide_id, pixel_location, testpixel, start_x, start_y, size, **kwargs
 ):
     setup_mock(kwargs)
     level = 0
     size_x = size
     size_y = size + 198
-    response = client.get(
-        f"/v1/slides/{slide_id}/region/level/{level}/start/{start_x}/{start_y}/size/{size_x}/{size_y}?image_format={image_format}&image_quality={image_quality}",
+    response = requests.get(
+        f"{wsi_service}/v1/slides/{slide_id}/region/level/{level}/start/{start_x}/{start_y}/size/{size_x}/{size_y}?image_format={image_format}&image_quality={image_quality}",
         stream=True,
     )
     assert response.status_code == 200
@@ -250,7 +264,7 @@ def test_get_slide_region_valid_brightfield(
     ],
 )
 def test_get_slide_region_valid_fluorescence(
-    client,
+    wsi_service,
     slide_id,
     channels,
     start_point,
@@ -264,8 +278,8 @@ def test_get_slide_region_valid_fluorescence(
 ):
     setup_mock(kwargs)
     level = 2
-    response = client.get(
-        f"/v1/slides/{slide_id}/region/level/{level}/start/{start_point[0]}/{start_point[1]}/size/{size[0]}/{size[1]}?image_format={image_format}&image_quality={image_quality}",
+    response = requests.get(
+        f"{wsi_service}/v1/slides/{slide_id}/region/level/{level}/start/{start_point[0]}/{start_point[1]}/size/{size[0]}/{size[1]}?image_format={image_format}&image_quality={image_quality}",
         stream=True,
     )
     assert response.status_code == 200
@@ -309,7 +323,7 @@ def test_get_slide_region_valid_fluorescence(
     ],
 )
 def test_get_slide_region_dedicated_channel(
-    client,
+    wsi_service,
     slide_id,
     level,
     channels,
@@ -324,8 +338,8 @@ def test_get_slide_region_dedicated_channel(
 ):
     setup_mock(kwargs)
     str_channels = "&".join([f"image_channels={str(ch)}" for ch in channels])
-    response = client.get(
-        f"/v1/slides/{slide_id}/region/level/{level}/start/{start_point[0]}/{start_point[1]}/size/{size[0]}/{size[1]}?image_format={image_format}&image_quality={image_quality}&{str_channels}",
+    response = requests.get(
+        f"{wsi_service}/v1/slides/{slide_id}/region/level/{level}/start/{start_point[0]}/{start_point[1]}/size/{size[0]}/{size[1]}?image_format={image_format}&image_quality={image_quality}&{str_channels}",
         stream=True,
     )
     assert response.status_code == 200
@@ -362,12 +376,11 @@ def test_get_slide_region_dedicated_channel(
         ("56ed11a2a9e95f87a1e466cf720ceffa", [5], 400),
     ],
 )
-def test_get_slide_region_invalid_channel(client, slide_id, channels, expected_response, **kwargs):
+def test_get_slide_region_invalid_channel(wsi_service, slide_id, channels, expected_response, **kwargs):
     setup_mock(kwargs)
     str_channels = "&".join([f"image_channels={str(ch)}" for ch in channels])
-    response = client.get(
-        f"/v1/slides/{slide_id}/region/level/2/start/0/0/size/64/64?{str_channels}",
-        stream=True,
+    response = requests.get(
+        f"{wsi_service}/v1/slides/{slide_id}/region/level/2/start/0/0/size/64/64?{str_channels}", stream=True
     )
     assert response.status_code == expected_response
 
@@ -381,15 +394,17 @@ def test_get_slide_region_invalid_channel(client, slide_id, channels, expected_r
         ("c801ce3d1de45f2996e6a07b2d449bca", (218, 217, 225), 15000, 15000, 30045),
         ("7304006194f8530b9e19df1310a3670f", (221, 170, 219), 50000, 90000, 30045),
         ("cdad4692405c556ca63185bee512e95e", (0, 0, 0), 30000, 30000, 30045),
+        ("5c1c0cc5cd3a501480fc6a4cb04ddda8", (0, 0, 0), 30000, 30000, 30045),
     ],
 )
-def test_get_slide_region_invalid(client, slide_id, testpixel, start_x, start_y, size, **kwargs):
+def test_get_slide_region_invalid(wsi_service, slide_id, testpixel, start_x, start_y, size, **kwargs):
     setup_mock(kwargs)
     level = 0
     size_x = size
     size_y = size + 198
-    response = client.get(
-        f"/v1/slides/{slide_id}/region/level/{level}/start/{start_x}/{start_y}/size/{size_x}/{size_y}", stream=True
+    response = requests.get(
+        f"{wsi_service}/v1/slides/{slide_id}/region/level/{level}/start/{start_x}/{start_y}/size/{size_x}/{size_y}",
+        stream=True,
     )
     assert response.status_code == 403
 
@@ -408,12 +423,15 @@ import timeit
         ("46061cfc30a65acab7a1ed644771a340", 1, 1, 2),
         ("cdad4692405c556ca63185bee512e95e", 1, 1, 5),
         ("c4682788c7e85d739ce043b3f6eaff70", 1, 1, 4),
+        ("5c1c0cc5cd3a501480fc6a4cb04ddda8", 1, 1, 4),
     ],
 )
-def test_get_slide_tile_timing(client, slide_id, tile_x, tile_y, level, **kwargs):
+def test_get_slide_tile_timing(wsi_service, slide_id, tile_x, tile_y, level, **kwargs):
     setup_mock(kwargs)
     tic = timeit.default_timer()
-    response = client.get(f"/v1/slides/{slide_id}/tile/level/{level}/tile/{tile_x}/{tile_y}", stream=True)
+    response = requests.get(
+        f"{wsi_service}/v1/slides/{slide_id}/tile/level/{level}/tile/{tile_x}/{tile_y}", stream=True
+    )
     assert response.status_code == 200
     get_image(response)
     toc = timeit.default_timer()
@@ -435,15 +453,16 @@ def test_get_slide_tile_timing(client, slide_id, tile_x, tile_y, level, **kwargs
         ("56ed11a2a9e95f87a1e466cf720ceffa", (30, 7, 6), 21, 22, (512, 512)),
         ("cdad4692405c556ca63185bee512e95e", (238, 238, 236), 210, 210, (256, 256)),
         ("c4682788c7e85d739ce043b3f6eaff70", (137, 75, 138), 210, 210, (256, 256)),
+        ("5c1c0cc5cd3a501480fc6a4cb04ddda8", (26, 26, 26), 21, 22, (128, 128)),
     ],
 )
 def test_get_slide_tile_valid(
-    client, image_format, image_quality, slide_id, testpixel, tile_x, tile_y, tile_size, **kwargs
+    wsi_service, image_format, image_quality, slide_id, testpixel, tile_x, tile_y, tile_size, **kwargs
 ):
     setup_mock(kwargs)
     level = 0
-    response = client.get(
-        f"/v1/slides/{slide_id}/tile/level/{level}/tile/{tile_x}/{tile_y}?image_format={image_format}&image_quality={image_quality}",
+    response = requests.get(
+        f"{wsi_service}/v1/slides/{slide_id}/tile/level/{level}/tile/{tile_x}/{tile_y}?image_format={image_format}&image_quality={image_quality}",
         stream=True,
     )
     assert response.status_code == 200
@@ -476,23 +495,25 @@ def test_get_slide_tile_valid(
         (10, 16, 422),  # level fails
     ],
 )
-def test_get_slide_tile_invalid(client, slide_id, tile_x, level, expected_response, **kwargs):
+def test_get_slide_tile_invalid(wsi_service, slide_id, tile_x, level, expected_response, **kwargs):
     setup_mock(kwargs)
-    response = client.get(f"/v1/slides/{slide_id}/tile/level/{level}/tile/{tile_x}/{tile_x}", stream=True)
+    response = requests.get(
+        f"{wsi_service}/v1/slides/{slide_id}/tile/level/{level}/tile/{tile_x}/{tile_x}", stream=True
+    )
     assert response.status_code == expected_response
 
 
 @requests_mock.Mocker(real_http=True, kw="requests_mock")
 @pytest.mark.parametrize("tile_size", [-1, 0, 1, 2500, 2501, 5000, 10000])
-def test_get_region_maximum_extent(client, tile_size, **kwargs):
+def test_get_region_maximum_extent(wsi_service, tile_size, **kwargs):
     wsi_settings = Settings()
     setup_mock(kwargs)
     level = 5
     start_x = 13
     start_y = 23
     slide_id = "7304006194f8530b9e19df1310a3670f"
-    response = client.get(
-        f"/v1/slides/{slide_id}/region/level/{level}/start/{start_x}/{start_y}/size/{tile_size}/{tile_size}",
+    response = requests.get(
+        f"{wsi_service}/v1/slides/{slide_id}/region/level/{level}/start/{start_x}/{start_y}/size/{tile_size}/{tile_size}",
         stream=True,
     )
     if tile_size * tile_size > wsi_settings.max_returned_region_size: