+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.2
+
+- Added support for Philips isyntax format
+- Restructured tests and test pipeline
+- Starting service via docker-compose
+
 ## 0.1.1
 
 - CI: Added version check